The prototype counter test bench now uses absoluate paths.
This is more elegant and does not require the hack used previously
where it had to cd into each primitives directory to compile with
Icarus; instead it hands Icarus absolute paths.

I've also added setUp and tearDown methods to the TestCase which I
hope to use in an intelligent manner to manage multiple tests. For
right now though the setUp method creates and stores the Cosimulation
object as an instance member such that the object can be used by other
methods (namely test_*). The tearDown method currently gets rid of the
executables created by Icarus after tests are done; this was needed to
avoid false positives when tests use older builds.

Note: the test bench Verilog file now no longer contains an `include
statement to load the DUT so if you're building with Icarus remember
to include that file. For example:

```bash
$ iverilog -o counter_tb counter_tb.v counter.v
```