The counter's MyHDL test-bench now uses the 'nose' testing framework.
This was chosen because it allows generator functions to be used to
run many tests at once and also because it is supported by setuptools.
On that note I have also added a 'setup.py' script which can be used
to install this package and also to run tests.

If you are a developer please use the 'develop' command to point Python
to your working copy so you don't have to constantly re-install the
package:

$ python setup.py develop

If you'd like to run all tests the preferred method is to use the
'test' command as follows:

$ python setup.py test

Or if you have the 'nose' package installed you can simply run
'nosetests' in the root directory (i.e. where the main package
folder is located).